+++ CHANGELOG.md
@@ -8,6 +8,9 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+### Changes
+
+- Changed: Separate Radio button into its own component.
 - Adds `TextInput` component to handle email, hidden, number, password, text, textarea, tel and url input types
 
 ## 0.21.2